requirements.txt
astroid==2.5.7
docstring-parser==0.8.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mikelint==1.0.3
Pillow==8.2.0
pylint==2.8.3
pytz==2021.1
PyYAML==5.4.1
toml==0.10.2
wrapt==1.12.1
